refactor(cmd): change pkg name of validator
